+++ src/position-resolver.ts
@@ -1,12 +1,6 @@
 import { Injectable, ElementRef } from '@angular/core';
 import { AxisResolver, AxisResolverFactory } from './axis-resolver';
-
-export interface PositionElements {
-  container: ElementRef;
-  documentElement: any;
-  isContainerWindow: boolean;
-  horizontal: boolean;
-}
+import { ContainerRef, PositionElements, PositionStats } from './models';
 
 @Injectable()
 export class PositionResolverFactory {
@@ -20,20 +14,47 @@ export class PositionResolverFactory {
 }
 
 export class PositionResolver {
+  private documentElement: ContainerRef;
+  private isContainerWindow: boolean;
+  public container: ContainerRef;
+
   constructor (private axis: AxisResolver, private options: PositionElements) {
+    this.resolveContainer(this.options.windowElement);
+    this.defineContainer(this.options.windowElement);
+  }
+
+  defineContainer(windowElement: ContainerRef) {
+    if (this.resolveContainer(windowElement)) {
+      this.container = windowElement;
+    } else {
+      this.container = windowElement.nativeElement;
+    }
+    return this.container;
+  }
+
+  resolveContainer(windowElement: ContainerRef): boolean {
+    const isContainerWindow = Object.prototype.toString.call(windowElement).includes('Window');
+    this.isContainerWindow = isContainerWindow;
+    return isContainerWindow;
+  }
+
+  getDocumentElement() {
+    return this.isContainerWindow
+      ? this.options.windowElement.document.documentElement
+      : null;
   }
 
   calculatePoints (element: ElementRef) {
-    return this.options.isContainerWindow
+    return this.isContainerWindow
       ? this.calculatePointsForWindow(element)
       : this.calculatePointsForElement(element);
   }
 
-  calculatePointsForWindow (element: ElementRef) {
+  calculatePointsForWindow (element: ElementRef): PositionStats {
     // container's height
-    const height = this.height(this.options.container);
+    const height = this.height(this.container);
     // scrolled until now / current y point
-    const scrolledUntilNow = height + this.pageYOffset(this.options.documentElement);
+    const scrolledUntilNow = height + this.pageYOffset(this.getDocumentElement());
     // total height / most bottom y point
     const totalToScroll = this.offsetTop(element.nativeElement) + this.height(element.nativeElement);
     return { height, scrolledUntilNow, totalToScroll };
@@ -42,17 +63,17 @@ export class PositionResolver {
   calculatePointsForElement (element: ElementRef) {
     let scrollTop    = this.axis.scrollTopKey();
     let scrollHeight = this.axis.scrollHeightKey();
+    const container = this.container;
 
-    const height = this.height(this.options.container);
+    const height = this.height(container);
     // perhaps use this.container.offsetTop instead of 'scrollTop'
-    const scrolledUntilNow = this.options.container[scrollTop];
+    const scrolledUntilNow = container[scrollTop];
     let containerTopOffset = 0;
-    const offsetTop = this.offsetTop(this.options.container);
+    const offsetTop = this.offsetTop(container);
     if (offsetTop !== void 0) {
       containerTopOffset = offsetTop;
     }
-    const totalToScroll = this.options.container[scrollHeight];
-    // const totalToScroll = this.offsetTop(this.$elementRef.nativeElement) - containerTopOffset + this.height(this.$elementRef.nativeElement);
+    const totalToScroll = container[scrollHeight];
     return { height, scrolledUntilNow, totalToScroll };
   }
 
@@ -62,7 +83,7 @@ export class PositionResolver {
 
     // elem = elem.nativeElement;
     if (isNaN(elem[offsetHeight])) {
-      return this.options.documentElement[clientHeight];
+      return this.getDocumentElement()[clientHeight];
     } else {
       return elem[offsetHeight];
     }
@@ -85,7 +106,7 @@ export class PositionResolver {
 
     // elem = elem.nativeElement;
     if (isNaN(window[pageYOffset])) {
-      return this.options.documentElement[scrollTop];
+      return this.getDocumentElement()[scrollTop];
     } else if (elem.ownerDocument) {
       return elem.ownerDocument.defaultView[pageYOffset];
     } else {