+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## v 0.2.9 
+## v 0.3.0 (2017/01/31) 
+
+### Updates
+* refactored infinite scroller to smaller modules with composition
+* added "models" - includes interfaces for development
+
+## v 0.2.9 (2017/01/13)
 
 ### Updates
 * added changelog.md